blocks: add multiply_const to its own folder
