+++ blocklib/blocks/include/gnuradio/blocks/meson.build
@@ -1,6 +1,4 @@
 blocks_headers = [
-
-    'multiply_const.hpp',
     'null_sink.hpp',
     'null_source.hpp',
     'vector_sink.hpp',
