+++ schedulers/mt/test/qa_scheduler_mt.cpp
@@ -57,8 +57,8 @@ TEST(SchedulerMTTest, MultiDomainBasic)
     }
 
     auto src = blocks::vector_source_f::make(input_data, false);
-    auto mult1 = blocks::multiply_const_ff::make(100.0);
-    auto mult2 = blocks::multiply_const_ff::make(200.0);
+    auto mult1 = blocks::multiply_const_ff::cpu(100.0);
+    auto mult2 = blocks::multiply_const_ff::cpu(200.0);
     auto snk = blocks::vector_sink_f::make();
 
     flowgraph_sptr fg(new flowgraph());
@@ -103,7 +103,7 @@ TEST(SchedulerMTTest, BlockFanout)
         std::vector<blocks::multiply_const_cc::sptr> mult_blks(nblocks);
 
         for (int i = 0; i < nblocks; i++) {
-            mult_blks[i] = blocks::multiply_const_cc::make(k, veclen);
+            mult_blks[i] = blocks::multiply_const_cc::cpu(k, veclen);
             sink_blks[i] = blocks::vector_sink_c::make();
         }
         flowgraph_sptr fg(new flowgraph());
